:construction::sparkles: Preparing for the implementation of Nuxt-og-image and implemented Nuxt SEO.
This commit prepares for the future implementation of Nuxt-og-image, which will be completed once the Nuxt team resolves an issue preventing the module from running normally on Nuxt version 3.8.2. Additionally, this commit introduces the implementation of Nuxt SEO, enhancing the website’s search engine optimization and visibility. These changes are part of our ongoing efforts to improve the functionality and user experience of our website.